refactor(modularization): Phase 4 Step 4-S2a — identity DB split (User + Role + OAuthAccount → aperag/domains/identity/db/models.py) (#25)
First sub-step of Phase 4 Step 4-S2 per @符炫炜 msg=d47fa490
Section 1 (DB split with G11 extension). Kept as a dedicated
sub-commit for CR manageability — Phase 3 Step 5b2* / 5b3 precedent.

**New module** ``aperag/domains/identity/db/models.py``:
- Class bodies of ``User`` + ``OAuthAccount`` + ``Role`` moved
  verbatim from ``aperag/db/models.py``.
- Local ``_random_id`` / ``_enum_column`` helpers (Phase 3
  per-domain DB module pattern — Phase 6 cleanup consolidates
  the helper twins onto ``aperag.db.base``).
- Docstring calls out G15 canonical: non-identity domains must
  never ``from aperag.domains.identity.db.models import Role``;
  they compare ``user.role == "admin"`` by literal against
  per-domain ``AuthenticatedUser`` / ``UserView`` Protocols.

**aperag/db/models.py**:
- Class bodies of ``User`` / ``OAuthAccount`` / ``Role``
  removed from the legacy aggregate, replaced by breadcrumb
  comments pointing at the new location.
- Re-export added via module-top import
  ``from aperag.domains.identity.db.models import OAuthAccount,
  Role, User``. Top import (vs. Phase 3's end-of-file shim) is
  required because the ``Invitation`` class still references
  ``EnumColumn(Role)`` in its body before the end-of-file
  Phase 3 shim block runs; this keeps the module body
  executable without reordering ``Invitation`` itself.

**Verification**:
- ``from aperag.db.models import User, Role, OAuthAccount,
  Invitation`` succeeds.
- ``Base.metadata`` still carries ``user`` / ``oauth_account`` /
  ``invitation`` tables (same three ORM class objects; no
  schema drift).
- Full unit suite: 671 passed / 29 skipped.
- Boundary + Phase 3 G11 re-export gates: 17/17 green.
- ``ruff format --check`` clean across 281 files.
